+++ lib/clickhouse-activerecord.rb
@@ -4,6 +4,7 @@ require 'active_record/connection_adapters/clickhouse_adapter'
 
 if defined?(Rails::Railtie)
   require 'clickhouse-activerecord/railtie'
+  require 'clickhouse-activerecord/schema'
   require 'clickhouse-activerecord/schema_dumper'
   require 'clickhouse-activerecord/tasks'
   ActiveRecord::Tasks::DatabaseTasks.register_task(/clickhouse/, "ClickhouseActiverecord::Tasks")