+++ lib/generators/clickhouse_migration_generator.rb
@@ -12,8 +12,8 @@ class ClickhouseMigrationGenerator < ActiveRecord::Generators::MigrationGenerato
   private
 
   def db_migrate_path
-    if defined?(Rails.application) && Rails.application
-      configured_migrate_path || default_migrate_path
+    if defined?(Rails.application) && Rails.application && respond_to?(:configured_migrate_path, true)
+      configured_migrate_path
     else
       default_migrate_path
     end